refactor: remove unused routines and rewrite loops
Next it would be a good idea to invent better data structures
for the log code than strings, or rewriting the struct-related code
like xfields, etc. Maybe rewriting the whole thing to use a stronger
approach now that we know how the compiler works better isn't the worst
idea for the future, but we don't have to do that right now, we can
still work with this at a high level with assumptions about the behavior.